feat: add dataProvider ts
